EpiSODIC: Epidemiological Signal Observation, Detection, Identification, and Classification

An automated surveillance system for infectious disease epidemiologists that detects aberrations in laboratory-confirmed infections, records epidemiologist's formal assessment with a full audit trail, and produces outbreak reports for clinical colleagues and infection prevention nurses. Applies evidence-based statistical methods such as Improved Farrington (2013) doi:10.1002/sim.5595 and 'EpiEstim' (Cori et al. 2013) doi:10.1093/aje/kwt133 . Ships with a synthetic demo data set requiring no external data or infrastructure.

See also

Author

Maintainer: Matthijs S. Berends m.berends@certe.nl

Authors:

Matthijs S. Berends m.berends@certe.nl

Other contributors:

Certe Medical Diagnostics & Advice Foundation (ROR) [copyright holder, funder]

University Medical Center Groningen (ROR) [copyright holder, funder]